+++ powerline_shell/utils.py
@@ -1,4 +1,5 @@
 import sys
+import os
 import threading
 
 py3 = sys.version_info[0] == 3
@@ -22,6 +23,7 @@ class RepoStats(object):
         'changed': u'\u270E',
         'new': u'?',
         'conflicted': u'\u273C',
+        'stash': u'\u2398',
         'git': u'\uE0A0',
         'hg': u'\u263F',
         'bzr': u'\u2B61\u20DF',
@@ -92,17 +94,19 @@ def warn(msg):
 
 
 class BasicSegment(object):
-    def __init__(self, powerline):
+    def __init__(self, powerline, segment_def):
         self.powerline = powerline
+        self.segment_def = segment_def  # type: dict
 
     def start(self):
         pass
 
 
 class ThreadedSegment(threading.Thread):
-    def __init__(self, powerline):
+    def __init__(self, powerline, segment_def):
         super(ThreadedSegment, self).__init__()
         self.powerline = powerline
+        self.segment_def = segment_def  # type: dict
 
 
 def import_file(module_name, path):
@@ -121,3 +125,28 @@ def import_file(module_name, path):
     else:
         import imp
         return imp.load_source(module_name, path)
+
+
+def get_PATH():
+    """Normally gets the PATH from the OS. This function exists to enable
+    easily mocking the PATH in tests.
+    """
+    return os.getenv("PATH")
+
+
+def get_subprocess_env(**envs):
+    defaults = {
+        # https://github.com/milkbikis/powerline-shell/pull/153
+        "PATH": get_PATH(),
+    }
+    defaults.update(envs)
+    env = dict(os.environ)
+    env.update(defaults)
+    return env
+
+
+def get_git_subprocess_env():
+    # LANG is specified to ensure git always uses a language we are expecting.
+    # Otherwise we may be unable to parse the output.
+    return get_subprocess_env(LANG="C")
+